+++ lib/foster_web/live/components/question_2.ex
